+++ CHANGELOG.md
@@ -9,6 +9,8 @@
   * logicapp
   * sqlserver
   * sqldatabase
+* Automatically update dependsOn attribute when resources that are added depend
+  on other resources that exist in the same template
 
 ## v0.1.2